refactored delete category and added error handling to delete dish service

+++ src/main/java/com/digitalmenu/menuservice/dish/DishService.java
@@ -32,6 +32,10 @@ public class DishService {
     }
 
     public void removeDish(Integer id) {
+        boolean exists = dishRepository.existsById(id);
+        if (!exists) {
+            throw new IllegalStateException("Dish with id " + id + " does not exists");
+        }
         dishRepository.deleteDishById(id);
     }
